ai_layer/requirements.txt
google-genai
kafka-python-ng
